third_party/WebKit/LayoutTests/media/video-volume-expected.txt

Test 'volume' attribute

EXPECTED (video.volume == '1') OK
RUN(video.volume = 0.5)
EXPECTED (video.volume == '0.5') OK
RUN(video.volume = Number.MIN_VALUE)
EXPECTED (video.volume == '5e-324') OK
RUN(video.volume = 0)
EXPECTED (video.volume == '0') OK
TEST(video.volume = 1.5) THROWS(DOMException.INDEX_SIZE_ERR: Index or size was negative, or greater than the allowed value.) OK
TEST(video.volume = -0.5) THROWS(DOMException.INDEX_SIZE_ERR: Index or size was negative, or greater than the allowed value.) OK
RUN(video.load())
EVENT(canplaythrough)
EXPECTED (video.volume == '0') OK
RUN(video.volume = 0.5)
EXPECTED (video.volume == '0.5') OK
TEST(video.volume = 1.5) THROWS(DOMException.INDEX_SIZE_ERR: Index or size was negative, or greater than the allowed value.) OK
TEST(video.volume = -0.5) THROWS(DOMException.INDEX_SIZE_ERR: Index or size was negative, or greater than the allowed value.) OK
END OF TEST
